+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fixed bug when vector format has no corresponding input value.  Was assigning
   an NA.  Now leaving the original value.
+* Made code compatible back to R version 3.6.0.
 * Added GitHub Actions for prior R version checks.
 
 # fmtr 1.5.0